# Projects/Project_2/apirest/static/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "PR2",
    "version": "1.0",
    "contact": {
      "name": "Marc Cervera",
      "email": "mcr34@udl.cat"
    },
    "description": "Project 2 Distrubuted Computing"
  },
  "servers": [
    {
      "url": "http://192.168.0.26:5000",
      "description": "Docker"
    }
  ],
  "paths": {
    "/home": {
      "parameters": [],
      "get": {
        "summary": "Get home",
        "tags": [],
        "responses": {
          "200": {
            "description": "User Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Home"
                },
                "examples": {}
              }
            }
          },
          "500": {
            "description": ""
          }
        },
        "operationId": "get-home",
        "description": "Retrieve the information of home",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "example": "casa 1"
            },
            "in": "query",
            "name": "description"
          }
        ]
      },
      "post": {
        "summary": "",
        "operationId": "post-home",
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  },
                  "x-examples": {
                    "Example 1": {
                      "message": "Home created "
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  },
                  "x-examples": {
                    "Example 1": {
                      "message": "Invalid data"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  },
                  "x-examples": {
                    "Example 1": {
                      "message": "Error creating home "
                    }
                  }
                }
              }
            }
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Home"
              },
              "examples": {
                "Example 1": {
                  "value": {
                    "name": "Casa",
                    "address": "Lleida",
                    "description": "casa 2",
                    "home_ownerid": 1
                  }
                },
                "Example 2": {
                  "value": {
                    "name": "Casa2",
                    "address": "Lleida",
                    "description": "casa 2"
                  }
                }
              }
            }
          }
        },
        "description": "Create a new home"
      },
      "put": {
        "summary": "",
        "operationId": "put-home",
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Home_update"
              },
              "examples": {
                "Example 1": {
                  "value": {
                    "name": "string",
                    "address": "string",
                    "description": "string",
                    "home_ownerid": 1,
                    "homeid": 1
                  }
                }
              }
            }
          }
        },
        "description": "Update data of an existing home"
      },
      "delete": {
        "summary": "",
        "operationId": "delete-home",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "description": "delete a home",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "home_ownerid": {
                    "type": "integer",
                    "example": 1,
                    "minimum": 1
                  },
                  "homeid": {
                    "type": "integer",
                    "example": 2
                  }
                },
                "required": [
                  "home_ownerid",
                  "homeid"
                ]
              },
              "examples": {
                "Example 1": {
                  "value": {
                    "home_ownerid": 1,
                    "homeid": 2
                  }
                }
              }
            }
          },
          "description": ""
        }
      }
    },
    "/sensor": {
      "post": {
        "summary": "Create New User",
        "operationId": "post-sensor",
        "responses": {
          "201": {
            "description": "User Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Sensor"
                },
                "examples": {}
              }
            }
          },
          "500": {
            "description": "Missing Required Information"
          }
        },
        "description": "Create a new sensor to a home",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Sensor"
              },
              "examples": {
                "Example 1": {
                  "value": {
                    "sensorID": "Temperatura",
                    "room": "Menjador",
                    "homeid": 1
                  }
                },
                "Example 2": {
                  "value": {
                    "sensorID": "Humitat",
                    "room": "Menjador",
                    "homeid": 1
                  }
                }
              }
            }
          },
          "description": ""
        }
      },
      "parameters": []
    },
    "/home/prts/{criteria}/{value}": {
      "get": {
        "summary": "Your GET endpoint",
        "tags": [],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "operationId": "get-home-prts-criteria-value",
        "description": ""
      },
      "parameters": [
        {
          "schema": {
            "type": "string",
            "example": "name"
          },
          "name": "criteria",
          "in": "path",
          "required": true
        },
        {
          "schema": {
            "type": "string",
            "example": "cas"
          },
          "name": "value",
          "in": "path",
          "required": true
        }
      ]
    },
    "/homeowner": {
      "post": {
        "summary": "Create a Home Owner",
        "operationId": "post-homeowner",
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Home_Owner"
              },
              "examples": {
                "example-1": {
                  "value": {
                    "name": "Marc Cervera2",
                    "password": "1234"
                  }
                }
              }
            }
          }
        }
      },
      "get": {
        "summary": "",
        "operationId": "get-homeowner",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "x-examples": {
                  "Example 1": {
                    "name": "Marc Cervera",
                    "password": "1234"
                  }
                },
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string"
                  }
                },
                "required": [
                  "name",
                  "password"
                ]
              },
              "examples": {
                "example-1": {
                  "value": {
                    "name": "Marc Cervera",
                    "password": "1234"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/home/{homeid}/sensor": {
      "parameters": [
        {
          "schema": {
            "type": "string",
            "example": "1"
          },
          "name": "homeid",
          "in": "path",
          "required": true
        }
      ],
      "get": {
        "summary": "Your GET endpoint",
        "tags": [],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "operationId": "get-home-homeid-sensor"
      }
    },
    "/homeowner/check/{home_ownerid}": {
      "parameters": [
        {
          "schema": {
            "type": "string"
          },
          "name": "home_ownerid",
          "in": "path",
          "required": true
        }
      ],
      "get": {
        "summary": "Your GET endpoint",
        "tags": [],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "operationId": "get-homeowner-home_ownerid"
      }
    }
  },
  "components": {
    "schemas": {
      "Home": {
        "description": "Home model",
        "type": "object",
        "x-examples": {
          "Example 1": {
            "name": "string",
            "address": "string",
            "description": "string",
            "home_ownerid": 0
          }
        },
        "title": "Home",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 100
          },
          "address": {
            "type": "string",
            "minLength": 1,
            "maxLength": 100
          },
          "description": {
            "type": "string",
            "minLength": 1
          },
          "home_ownerid": {
            "type": "integer",
            "default": 1,
            "minimum": 1,
            "example": 1
          }
        },
        "required": [
          "name",
          "address",
          "description"
        ]
      },
      "Sensor": {
        "type": "object",
        "x-examples": {
          "Example 1": {
            "sensorID": "string",
            "room": "string",
            "homeid": 1
          }
        },
        "properties": {
          "sensorID": {
            "type": "string",
            "minLength": 1,
            "example": "Temperatura"
          },
          "room": {
            "type": "string",
            "minLength": 1
          },
          "homeid": {
            "type": "integer",
            "example": 1
          }
        },
        "required": [
          "sensorID",
          "room",
          "homeid"
        ]
      },
      "Home_Owner": {
        "description": "Home_Owner model",
        "type": "object",
        "x-examples": {
          "Example 1": {
            "name": "string",
            "password": "string",
            "home_ownerid": 1
          }
        },
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 50
          },
          "password": {
            "type": "string",
            "minLength": 1,
            "maxLength": 50
          },
          "home_ownerid": {
            "type": "integer",
            "minimum": 1
          }
        },
        "required": [
          "name",
          "password"
        ]
      },
      "Home_update": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "address": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "home_ownerid": {
            "type": "integer"
          },
          "homeid": {
            "type": "integer"
          }
        },
        "x-examples": {
          "Example 1": {
            "name": "string",
            "address": "string",
            "description": "string",
            "home_ownerid": 0,
            "homeid": 1
          }
        },
        "title": "Home_update",
        "x-stoplight": {
          "id": "8985e66762a1b"
        }
      }
    }
  }
}